+++ frontend/src/components/admin/LayoutDashboard.vue
@@ -127,6 +127,7 @@ const dashBoard = ref({
         redirect_links_count: 0,
         roles_count: 0,
         j4r_servers_count: 0,
+        image_reports_count: 0,
     },
 });